feat(search): show call timestamp on call-record content hit rows
