+++ js/app/packages/block-call/component/transcript-playback.ts
@@ -1,5 +1,7 @@
 import type { CallRecordTranscriptSegment } from '@service-storage/generated/schemas/callRecordTranscriptSegment';
 
+export { formatVideoTimestamp } from '@core/util/duration';
+
 export function sortTranscriptSegments(
   transcript: CallRecordTranscriptSegment[]
 ): CallRecordTranscriptSegment[] {
@@ -41,17 +43,6 @@ export function getActiveTranscriptSequenceNum(
   }
 
   return activeSequenceNum;
-}
-
-export function formatVideoTimestamp(totalSeconds: number): string {
-  const clamped = Math.max(0, Math.floor(totalSeconds));
-  const hours = Math.floor(clamped / 3600);
-  const minutes = Math.floor((clamped % 3600) / 60);
-  const seconds = clamped % 60;
-  if (hours > 0) {
-    return `${hours}:${String(minutes).padStart(2, '0')}:${String(seconds).padStart(2, '0')}`;
-  }
-  return `${minutes}:${String(seconds).padStart(2, '0')}`;
 }
 
 export function getSegmentVideoSeconds(
