+++ js/app/packages/core/util/duration.ts
@@ -1,0 +1,14 @@
+/**
+ * Formats a duration as `m:ss` (or `h:mm:ss` past one hour). Used for
+ * video / call playback offsets.
+ */
+export function formatVideoTimestamp(totalSeconds: number): string {
+  const clamped = Math.max(0, Math.floor(totalSeconds));
+  const hours = Math.floor(clamped / 3600);
+  const minutes = Math.floor((clamped % 3600) / 60);
+  const seconds = clamped % 60;
+  if (hours > 0) {
+    return `${hours}:${String(minutes).padStart(2, '0')}:${String(seconds).padStart(2, '0')}`;
+  }
+  return `${minutes}:${String(seconds).padStart(2, '0')}`;
+}
